Refactor: Move petSlice into a better location
