+++ web/src/app/store.js
@@ -1,7 +1,7 @@
 import { applyMiddleware, configureStore } from '@reduxjs/toolkit';
 import thunkMiddleware from 'redux-thunk';
 
-import petReducer from '../features/pets/petSlice';
+import petReducer from '../components/petSlice';
 
 export const setupStore = (preloadedState) => {
   return configureStore(
