+++ web/src/components/pet.d.ts
